add papers

+++ tinygrad/apps/llm2.py
@@ -51,14 +51,14 @@ def apply_rope(x:Tensor, start_pos:int|UOp, base:int=150_000, scale:float=32.0, 
   def rotate(x_pairs, freqs):
     angles = ((Tensor.arange(T, dtype="float32") + start_pos)[:, None] * freqs[None, :]).reshape(1, 1, T, half)
     # contiguous here allows RoPE to be pruned in the JIT
-    cos, sin = angles.cos().cast('float').contiguous(), angles.sin().cast('float').contiguous()
+    cos, sin = angles.cos().cast(x_pairs.dtype).contiguous(), angles.sin().cast(x_pairs.dtype).contiguous() # todo: cast to float32 ??
     return Tensor.stack(x_pairs[..., 0] * cos - x_pairs[..., 1] * sin, x_pairs[..., 0] * sin + x_pairs[..., 1] * cos, dim=-1)
 
-  # rope
+  # rope https://arxiv.org/pdf/2104.09864
   if scale <= 1:
     return rotate(x.reshape(B, H, T, 2, half).transpose(-1, -2), freqs).transpose(-1, -2).reshape(B, H, T, Hd)
 
-  # yarn
+  # yarn https://arxiv.org/pdf/2309.00071
   def _ratio(ntk): return half * math.log(initial_context_length / (ntk * 2 * math.pi)) / math.log(base)
   x = x * (0.1 * math.log(scale) + 1.0)
   low, high = _ratio(ntk_beta), _ratio(ntk_alpha)
